+++ lib/rubocop/cop/style/empty_else.rb
@@ -4,11 +4,72 @@ module RuboCop
   module Cop
     module Style
       # Checks for empty else-clauses, possibly including comments and/or an
-      # explicit `nil`.
+      # explicit `nil` depending on the EnforcedStyle.
+      #
+      # SupportedStyles:
+      #
+      # @example
+      #   # good for all styles
+      #   if condition
+      #     statement
+      #   else
+      #     statement
+      #   end
+      #
+      #   # good for all styles
+      #   if condition
+      #     statement
+      #   end
+      #
+      # empty - warn only on empty else
+      #   @example
+      #   # bad
+      #   if condition
+      #     statement
+      #   else
+      #   end
+      #
+      #   # good
+      #   if condition
+      #     statement
+      #   else
+      #     nil
+      #   end
+      #
+      # nil - warn on else with nil in it
+      #   @example
+      #   # bad
+      #   if condition
+      #     statement
+      #   else
+      #     nil
+      #   end
+      #
+      #   # good
+      #   if condition
+      #     statement
+      #   else
+      #   end
+      #
+      # both - warn on empty else and else with nil in it
+      #   @example
+      #   # bad
+      #   if condition
+      #     statement
+      #   else
+      #     nil
+      #   end
+      #
+      #   # bad
+      #   if condition
+      #     statement
+      #   else
+      #   end
       class EmptyElse < Cop
         include OnNormalIfUnless
+        include ConfigurableEnforcedStyle
 
-        MSG = 'Redundant empty `else`-clause.'
+        MSG = 'Redundant `else`-clause.'
 
         def on_normal_if_unless(node)
           check(node, if_else_clause(node))
@@ -21,25 +82,33 @@ module RuboCop
         private
 
         def check(node, else_clause)
-          return unless node.loc.else
-          return if else_clause && else_clause.type != :nil
-
-          add_offense(node, :else, MSG)
-        end
-
-        def if_else_clause(node)
-          keyword = node.loc.keyword
-          if keyword.is?('if')
-            node.children[2]
-          elsif keyword.is?('elsif')
-            node.children[2]
-          elsif keyword.is?('unless')
-            node.children[1]
+          case style
+          when :empty
+            empty_check(node, else_clause)
+          when :nil
+            nil_check(node, else_clause)
+          when :both
+            both_check(node, else_clause)
           end
         end
 
-        def case_else_clause(node)
-          node.children.last
+        def empty_check(node, else_clause)
+          add_offense(node, :else, MSG) if node.loc.else && else_clause.nil?
+        end
+
+        def nil_check(node, else_clause)
+          return unless else_clause && else_clause.type == :nil
+          add_offense(node, node.location, MSG)
+        end
+
+        def both_check(node, else_clause)
+          return if node.loc.else.nil?
+
+          if else_clause.nil?
+            add_offense(node, :else, MSG)
+          elsif else_clause.type == :nil
+            add_offense(node, :else, MSG)
+          end
         end
       end
     end